New Session method prepare_request_for_redirect.
This is almost entirely moving code around.  The work done by the new
method was formerly embedded in `resolve_redirects`, which now contains
only the main loop.

Since the new method only takes one argument, the previous response,
the footgun where you could pass the wrong request object as the second
argument to `resolve_redirects` is eliminated; take the opportunity to
further rectify the `resolve_redirects` calling convention (it will now
accept and forward arbitrary kwargs).  This allows the guts of `Session.send`
to be streamlined a little.